Re-solve the production network before the injection networks
An injection group target (GCONINJE VREP/REIN) is a function of the produced
voidage, so within a network sub-iteration the injectors must see this
iteration's production, not the previous one's. The well loop ran in deck
order, so that was a matter of luck.

Producers are now re-solved first, then the injectors. The group data is
refreshed in between only when a deck has both a production and an injection
network -- the only case where the producers re-solved here can feed an
injection target in the same sub-iteration -- so nothing else pays for the
extra update. details::activeNetworks() already returns the production
network first, so the pressure computation was in the right order already.

No deck in opm-tests has both network kinds, so this is a correctness guard
rather than a fix for an observed failure: NETWORK-01, NETWORK-01_STANDARD,
NETWORK-01-WTEST, GNETINJE_GAS-01 and GNETINJE_WAT-01 are all bit-identical
(compareECL -t SMRY at 1e-9) before and after.
